API: Unify return values for overcurrent protection functions

+++ app/src/bms.h
@@ -286,26 +286,38 @@ int bms_apply_temp_limits(BmsConfig *conf);
 /**
  * Apply discharge short circuit protection (SCP) current threshold and delay.
  *
+ * If the setpoint does not exactly match a possible setting in the BMS IC, it is rounded to the
+ * closest allowed value and this value is written back to the BMS config.
+ *
  * @param conf BMS configuration containing the limit settings
- * @returns applied current threshold value or 0 in case of error
+ *
+ * @returns 0 on success, otherwise negative error code.
  */
-float bms_apply_dis_scp(BmsConfig *conf);
+int bms_apply_dis_scp(BmsConfig *conf);
 
 /**
  * Apply discharge overcurrent protection (OCP) threshold and delay.
  *
+ * If the setpoint does not exactly match a possible setting in the BMS IC, it is rounded to the
+ * closest allowed value and this value is written back to the BMS config.
+ *
  * @param conf BMS configuration containing the limit settings
- * @returns applied current threshold value or 0 in case of error
+ *
+ * @returns 0 on success, otherwise negative error code.
  */
-float bms_apply_dis_ocp(BmsConfig *conf);
+int bms_apply_dis_ocp(BmsConfig *conf);
 
 /**
  * Apply charge overcurrent protection (OCP) threshold and delay.
  *
+ * If the setpoint does not exactly match a possible setting in the BMS IC, it is rounded to the
+ * closest allowed value and this value is written back to the BMS config.
+ *
  * @param conf BMS configuration containing the limit settings
- * @returns applied current threshold value or 0 in case of error
+ *
+ * @returns 0 on success, otherwise negative error code.
  */
-float bms_apply_chg_ocp(BmsConfig *conf);
+int bms_apply_chg_ocp(BmsConfig *conf);
 
 /**
  * Apply cell undervoltage protection (UVP) threshold and delay.
